backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.9.2
pydantic-settings==2.4.0
pyrogram==2.0.106
tgcrypto==1.2.5
cryptography==42.0.7
SQLAlchemy==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7

